+++ src/engine/display/display.cpp
@@ -1,4 +1,4 @@
-#include "display.h"
+#include "../display.h"
 
 #include <windows.h>
